+++ CHANGELOG.md
@@ -22,12 +22,15 @@
   verifying ([#126]).
 - **[IMPROVED]** Added checks for unused stub criteria ([#126]).
 - **[FIXED]** Fixed bug when mocking traits with magic call methods ([#127]).
+- **[FIXED]** Ad-hoc mocks that differ only by function body no longer result in
+  re-use of the same mock class ([#131]).
 
 [#98]: https://github.com/eloquent/phony/issues/98
 [#117]: https://github.com/eloquent/phony/issues/117
 [#123]: https://github.com/eloquent/phony/issues/123
 [#126]: https://github.com/eloquent/phony/issues/126
 [#127]: https://github.com/eloquent/phony/issues/127
+[#131]: https://github.com/eloquent/phony/issues/131
 
 ## 0.7.0 (2015-12-20)